requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
statsmodels>=0.13.0
scikit-learn>=0.24.0
openai>=0.27.0
python-dotenv>=0.19.0
yfinance>=0.1.70 